refactor: update code
